requirements.txt
streamlit
google-generativeai
streamlit-lottie